renamed note.rs into pitch.rs
